removed unused files from list; Added test for dotfile existance

+++ makesymlinks.sh
@@ -9,7 +9,7 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="bashrc vimrc vim zshrc oh-my-zsh private scrotwm.conf Xresources"    # list of files/folders to symlink in homedir
+files="bashrc vimrc vim"    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -25,9 +25,12 @@ echo "done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ -e ~/.$file ]
+    then
+      echo "Moving .$file from ~ to $olddir"
+      mv ~/.$file $olddir/
+      echo "Creating symlink to $file in home directory."
+      ln -s $dir/$file ~/.$file
+    fi
 done
